+++ src/main/java/vendingmachine/domain/VendingMachine.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import vendingmachine.Coin;
 import vendingmachine.domain.coin.VendingMachineCoin;
 import vendingmachine.util.MoneyGenerator;
