+++ src/main/java/vendingmachine/domain/coin/VendingMachineCoin.java
@@ -3,7 +3,7 @@ package vendingmachine.domain.coin;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
-import vendingmachine.domain.Coin;
+import vendingmachine.Coin;
 
 public class VendingMachineCoin {
